Added configuration options default_epsg and quiet
If 'default_epsg' is set, PostGIS geometries whose SRID is not set will be
set to the default EPSG for further processing.

If 'quiet' is set, warnings will be disabled. For example using a 'default_epsg'
will cause a warning unless warnings are disabled.

Use Spine::Exception::Trace when possible