Added CoordinateTransformationCache (BRAINSTORM-1857)
OGRCoordinateTransformation is not thread safe, hence a cached copy cannot be used in multiple threads.
CoordinateTransformationCache is really a pool of objects, and may contain as many instances of a single
OGRCoordinateTransformation as seems to be needed by the running program - a new copy is always created
if one is not available when needed.

The implementation could be different if the newer GDAL 3.0 provided OGRCoordinateTransformation::Clone()
was available right now, but it is probable that this version would be faster than calling Clone() to create
a instance for the client instead of just returning an available instance.